Pathfinding Edge bug
The bug was caused from the pathfinding not being able to access the target node as that was being blocked by a wall. Added code that during these scenarios the blocked node will instead use one of its not blocked neighbours instead.

Also added optimisation for pathfinding calls. Only does the call on player node change.