Bug fixing + QOL
So the optimisation I did had one small issue: It wouldn't update it's direction along the path if the target didn't move. Added another check so it knows if its node has also changed instead of only the targets.

Also changed the gizmo of the leap to accurately represent the new logic of the leap.