Pathfinding leaping and wall collisions fixed
Okay so few things, I misinterpreted the get neighbour node for getting unblocked nodes only and not all neighbour nodes in the grid so needed to update some logic.

Updated the leap logic and added a new function in the pathfinding manager for finding the nearest node that is unblocked from the target in a specified direction.

Also removed all Debug.Log comments as it will make it easier to sweep debug statements in the future.