Cursor fixes
Turns out the build in cursor doesn't scale with screen size. In most cases with a cursor this is wanted but we have a reticle type cursor so this is not what we want.

The cursor has been changed to a UI object and now scales with screen size and draws over all other UI objects.